move files and fix imports for tfrecord generation scripts
