+++ training/.gitignore
@@ -1,0 +1,3 @@
+dev_accs
+pred_results
+example_data
