+++ training/upload_devaccs.sh
